refactor: decompose provider sign-in orchestration
